Remove Filesystem interface
